[refactor] separate types files
