+++ src/types/index.ts
@@ -1,0 +1,6 @@
+// api
+export interface KAKAOParams {
+  query: string;
+  sort: string;
+  size: number;
+}
